% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_reference}
\alias{align_reference}
\title{Filter and align single or paired-end reads to a reference genome.}
\usage{
align_reference(
  RA_fastqs,
  RB_fastqs = NULL,
  reference,
  mapQ = 5,
  remove_duplicates = TRUE,
  remove_improper_pairs = TRUE,
  par = 1
)
}
\arguments{
\item{RA_fastqs}{character. Vector of filepaths for the RA (read one) files.}

\item{RB_fastqs}{character or NULL, default NULL. Vector of filepaths for the
RB (read one) files.}

\item{reference}{character. Filepath for the reference genome to use for the
alignment.}

\item{mapQ}{numeric, default 5. Reads with mapping qualities lower than mapQ
will be removed using \code{samtools view -q mapQ}. Only valid for
paired-end reads.}

\item{remove_duplicates}{logical, default TRUE. If TRUE, PCR duplicates
(clones) will be removed using \code{samtools markdup}. Only valid for
paired-end reads.}

\item{remove_improper_pairs}{logical, default TRUE. If TRUE, improperly
paired reads reads will be removed using \code{samtools view -f 0x2}. Only
valid for paired-end reads.}

\item{par}{numeric, default 1. Number of cores to use for the alignments.}
}
\value{
Generates 'x.sort.flt.bam' and 'x.sort.flt.bam.bai' files (sorted
  alignments and index files, respectively), where 'x' is the prefix for the
  RA. fastqs. In R, returns a vector of output file paths.
}
\description{
Aligns split, single- or paired-end reads to a reference genome using the bwa
mem algorithm, them filters the resulting alignments using a few different
approaches via samtools.
}
\details{
Paired-end reads are filtered following alignment by first removing PCR
duplicates using \code{samtools fixmate} and \code{samtools markdup}, then
poorly mapped reads using \code{samtools view} with \code{-q 5}, and then
improperly paired reads using \code{samtools view} with  \code{-f 0x2}. The
full alignment script can be accessed using \code{system.file("shell",
"run_align.sh", "alignR")}.

Single-end reads are not filtered following alignment, although filtering can
(and usually should) be done while calling genotypes or doing other
down-stream analyses. The full alignment script can be accessed using
\code{system.file("shell", "run_align_single.sh", "alignR")}.
}
\author{
William Hemstrom
}
