% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_to_reference}
\alias{align_to_reference}
\title{Filter and align paired-end reads to a reference genome.}
\usage{
align_to_reference(RA_fastqs, RB_fastqs, reference, par = 1)
}
\description{
Filter and align paired-end reads to a reference genome.
}
