% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignR-package.R
\docType{package}
\name{alignR-package}
\alias{alignR}
\alias{alignR-package}
\title{alignR: Landmark digitizer for alignment of surface meshes}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
