% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignR.R
\name{alignR}
\alias{alignR}
\title{Landmark alignR}
\usage{
alignR(file_dir, file_name = "Landmarks.txt", loadAll = TRUE)
}
\arguments{
\item{file_dir}{A directory containing surface files to be analyzed}

\item{file_name}{A name for the landmark data file to be generated}

\item{loadAll}{A logical value which determines whether all surface files should be loaded into a list or if they should be loaded only as needed for digitizing}
}
\description{
Shiny application to collect 3D landmark data on surface files
}
\details{
This function prepares mesh object data for landmarking and loads the landmarking application.
Using a combination of rgl, Rvcg, and shiny functions, meshes are loaded into an rglwidget which can
be interactively manipulated. Some compromises still exist, but this interface enables easier
switching between specimens and re-landmarking, as landmarks can be placed in any order and
can be loaded for later reassessment.

There are three options for type of landmarking: 1. fixed / discrete landmarks, 2. a mix of discrete
and surface pseudolandmarks, & 3. alignment landmarks for fitting surface pseudolandmarks. Switching
between analysis type is done interactively in the open application.
}
