% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gui.R
\name{alignR_gui}
\alias{alignR_gui}
\title{Run the alignR GUI}
\usage{
alignR_gui()
}
\description{
Runs \code{alignR}'s Graphical User Interface (GUI) \code{\link[shiny]{shiny}}
App to help simplify demultiplexing, alignment, and genotyping using the full
suite of tools available in the package.
}
\author{
William Hemstrom
}
