% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_Functions.R
\name{checkLMs}
\alias{checkLMs}
\title{Check Landmarks}
\usage{
checkLMs(current_lm, check)
}
\arguments{
\item{lm_data}{An array of landmark data.}

\item{current_sp}{The current specimen, referencing the reactive object cur_sp().}

\item{lm_labels}{The landmark labels, referencing the reactive object LM_values().}
}
\value{
landmarks An array of loaded landmark data.
}
\description{
Checking of new landmark position.
}
\details{
An internal function for loading the most recently submitted landmark for the user to verify.
Used by submitLM() in the shiny server.
}
\keyword{internal}
