% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_Functions.R
\name{clearLMs}
\alias{clearLMs}
\title{Clear Landmarks}
\usage{
clearLMs(lm_labels)
}
\arguments{
\item{lm_labels}{The landmark labels, referencing the reactive object LM_values().}
}
\value{
landmarks An blank array of landmarks.
}
\description{
Clear current landmark data for current specimen.
}
\details{
An internal function for clearing the current specimen's landmark data. Used by clearLM()
in the shiny server.
}
\keyword{internal}
