% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATK.R
\name{concat_vcfs}
\alias{concat_vcfs}
\title{Concatenate ".vcf" files.}
\usage{
concat_vcfs(
  vcfs,
  outfile = paste0(tools::file_path_sans_ext(vcfs[1]), "_concat.vcf")
)
}
\arguments{
\item{vcfs}{Character, vector of "vcf" files.}

\item{outfile}{Character,
default \code{paste0(tools::file_path_sans_ext(vcfs[1], "_concat.vcf"))}.
Name for final, concatenated "vcf" file.}
}
\description{
Concatenates (joins) ".vcf" files with different loci but identical individuals
into a single ".vcf" file.
}
