% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{downsample_bams}
\alias{downsample_bams}
\title{Down sample mapped reads across multiple bam files.}
\usage{
downsample_bams(
  bams,
  reference,
  low_cut = 0,
  high_cut = Inf,
  tabulate_only = FALSE
)
}
\arguments{
\item{bams}{character. Vector of bamfile filepaths. If files are not indexed,
they will be beforehand.}

\item{reference}{character or numeric. Path to reference genome to which
bamfiles were mapped OR the length of the genome/sequenced region, in bp.}

\item{low_cut}{numeric, default 0. Lower bound below which bam files should
be dropped from further analysis.}

\item{high_cut}{numeric, default Inf. Upper bound; bam files with higher
coverage will be downsampled to this coverage.}

\item{tabulate_only}{logical, default FALSE. If TRUE, no downsampling is
performed; instead, a table of counts, depths, subsample levels is
returned. Useful to run prior to an actual downsampling run to determine
\code{high_cut} and \code{low_cut}.}
}
\value{
Generates files downsampled to requested level. In R, returns a list
  of file paths to subset files OR, if \code{tabulate_only} is TRUE, a table
  of sample depth statistics.
}
\description{
Randomly downsample bam files to a target value \code{high_cut}. Optionally
also detect any files below \code{low_cut}. Files between the \code{high_cut}
and \code{low_cut} will be copied without downsampling. Return a list of
downsampled files without anything below \code{low_cut}. Duplicate removal
is strongly recommended before downsampling.
}
\author{
William Hemstrom
}
