% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{equalize_read_counts}
\alias{equalize_read_counts}
\title{Equalize the number of reads across multiple fastq or fastq files.}
\usage{
equalize_read_counts(
  RA_fastqs,
  RB_fastqs = NULL,
  nreads = "lowest",
  outfile_prefix = "subset_"
)
}
\arguments{
\item{RA_fastqs}{character. Vector of filepaths for the RA (read one) files.}

\item{RB_fastqs}{character or NULL, default NULL. Vector of filepaths for the
RB (read one) files. If NULL, assumes reads are single-end.}

\item{nreads}{numeric or "lowest", default "lowest". Number of reads to keep.
If "lowest", will keep n reads, where n is the number of reads in the
smallest .fastq/.fasta file.}

\item{outfile_prefix}{character. Prefix to be appended to each resulting
.fastq/.fastq file.}
}
\value{
Generates files subset to the specified size. In R, returns a list of
  RA (and RB if paired-end) file paths to subset files.
}
\description{
Subset out the first n reads of every fastq/fastq file, where n is either the
smallest number of reads across all samples or user supplied.
}
\author{
William Hemstrom
}
