% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype_bams}
\alias{genotype_bams}
\title{Extract SNP genotypes from aligned bamfiles using ANGSD.}
\usage{
genotype_bams(
  bamfiles,
  outfile_prefix = "genotypes",
  minInd = floor(length(bamfiles)/2),
  genotyper = "SAMtools",
  SNP_pval = 0.00000001,
  doGeno = "NN",
  postCutoff = 0.95,
  minQ = 20,
  minMapQ = 20,
  unzip = FALSE,
  par = 1
)
}
\arguments{
\item{bamfiles}{character. Vector of filepaths to the bamfiles containing
data for the individuals to genotype.}

\item{outfile_prefix}{character, default "genotypes". Prefix for the
outfiles.}

\item{minInd}{numeric, default \code{floor(length(bamfiles)/2)}. Minimum
number of individuals a locus must be sequenced in in order to call
genotypes.}

\item{genotyper}{character, default "SAMtools". Name of the genotyper to use.
Options: \itemize{\item{SAMtools} \item{GATK} \item{SOAPsnp} \item{phys}
\item{sample: } randomly sample reads to make the genotypes (therefore
posteriors are either 0, .5, or 1).} See ANGSD documentation for details.}

\item{SNP_pval}{numeric, default 0.00000001. Minimum p-value that must be met
\emph{for the probability that a given locus is a SNP} for a locus to be
genotyped.}

\item{doGeno}{character, default "NN". Vector of styles for which genotypes
will be printed. Options: \itemize{\item{major_minor: } Print the major and
minor alleles. \item{numeric: } Print alleles as 0, 1, 2, or -1 for the
homozygous major, hetoerzygote, homozygous minor, or missing data,
respectively. \item{NN: } Print alleles as paired nucleotides (AA, AC, CC,
for example). NN denotes missing data. \item{all_posteriors: } Prints
posterior probabilities for \emph{all three} genotypes at each locus.
\item{called_posterior: } Prints the posterior probability for \emph{only
the called genotype}. \item{binary: } Prints the posterior probabilities
for the three genotypes \emph{in binary}. This will convert all other
selected options to binary as well.} If multiple options are selected, they
\emph{all be printed to a single file}, in the order provided in this
documentation.}

\item{postCutoff}{numeric, default 0.95. The minimum posterior probability
required for the most likely genotype for the genotype to be printed at for
a given individual at a given locus.}

\item{minQ}{numeric, default 20. The minimum Phred \emph{sequencing} quality
score needed for a given sequenced base on a single read to be considered
during genotyping. The default, 20, corresponds to 99\% accuracy.}

\item{minMapQ}{numeric, default 20. The minimum Phred \emph{mapping} quality
score needed for a given sequenced base on a single read to be considered
during genotyping. The default, 20, corresponds to 99\% accuracy.}

\item{unzip}{logical, default FALSE. If TRUE, resulting genotype file will be
unzipped.}

\item{par}{numeric, default 1. Number of cores to allow ANGSD to use for
genotyping.}
}
\value{
Generates genotype files in the requested format named
  outfile_prefix.geno(.gz) in the directory of the first bamfile. The file
  will be zipped (.gz) if \code{unzip} is not TRUE. In R, returns file path
  to genotype output.
}
\description{
Uses the ANGSD (Analysis of Next Gen Sequencing Data) software to genotype
individuals across a set of bamfiles using several different filtering
options or genotyping methods. This function is a \emph{wrapper for ANGSD}.
}
\author{
William Hemstrom

Michael Miller
}
