% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype_bams_ANGSD}
\alias{genotype_bams_ANGSD}
\title{Extract SNP genotypes from aligned bamfiles using ANGSD.}
\usage{
genotype_bams_ANGSD(
  bamfiles,
  outfile_prefix = "genotypes",
  minInd = floor(length(bamfiles)/2),
  genotyper = "GATK",
  SNP_pval = 1e-08,
  doGeno = "NN",
  postCutoff = 0.95,
  minQ = 20,
  minMapQ = 20,
  minMaf = 0,
  unzip = FALSE,
  doVcf = FALSE,
  filter_paralogs = FALSE,
  filter_paralogs_alpha = 0.0016,
  filter_paralogs_buffer = 1000,
  filter_paralogs_populations = rep("only_pop", length(bamfiles)),
  filter_paralogs_reference = FALSE,
  rf = FALSE,
  par = 1
)
}
\arguments{
\item{bamfiles}{character. Vector of filepaths to the bamfiles containing
data for the individuals to genotype.}

\item{outfile_prefix}{character, default "genotypes". Prefix for the
outfiles.}

\item{minInd}{numeric, default \code{floor(length(bamfiles)/2)}. Minimum
number of individuals a locus must be sequenced in in order to call
genotypes.}

\item{genotyper}{character, default "GATK". Name of the genotyper to use.
Options: \itemize{\item{SAMtools} \item{GATK} \item{SOAPsnp} \item{phys}
\item{sample: } randomly sample reads to make the genotypes (therefore
posteriors are either 0, .5, or 1).} See ANGSD documentation for details.}

\item{SNP_pval}{numeric, default 0.00000001. Minimum p-value that must be met
\emph{for the probability that a given locus is a SNP} for a locus to be
genotyped.}

\item{doGeno}{character, default "NN". Format in which genotypes will be
printed. Options: \itemize{\item{major_minor: } Print the major and minor
alleles. \item{numeric: } Print alleles as 0, 1, 2, or -1 for the
homozygous major, hetoerzygote, homozygous minor, or missing data,
respectively. \item{NN: } Print alleles as paired nucleotides (AA, AC, CC,
for example). NN denotes missing data. \item{all_posteriors: } Prints
posterior probabilities for \emph{all three} genotypes at each locus.
\item{called_posterior: } Prints the posterior probability for \emph{only
the called genotype}. \item{binary: } Prints the posterior probabilities
for the three genotypes \emph{in binary}.}}

\item{postCutoff}{numeric, default 0.95. The minimum posterior probability
required for the most likely genotype for the genotype to be printed at for
a given individual at a given locus.}

\item{minQ}{numeric, default 20. The minimum Phred \emph{sequencing} quality
score needed for a given sequenced base on a single read to be considered
during genotyping. The default, 20, corresponds to 99\% accuracy.}

\item{minMapQ}{numeric, default 20. The minimum Phred \emph{mapping} quality
score needed for a given sequenced base on a single read to be considered
during genotyping. The default, 20, corresponds to 99\% accuracy.}

\item{minMaf}{numeric, default 0. The minimum Minor Allele Frequency (MAF)
needed to keep a locus.}

\item{unzip}{logical, default FALSE. If TRUE, resulting genotype file will be
unzipped.}

\item{doVcf}{logical, default FALSE. If TRUE, create a vcf file. Not
supported for \code{doGeno = 'major_minor'}. If \code{doGeno} is
\code{numeric} or \code{NN}, done using an 'in-house' perl script.
Otherwise done using the \code{-doBcf} option in ANGSD, then converted to a
.vcf using \code{bcftools}.}

\item{filter_paralogs}{logical, default FALSE. If true \code{ngsParalogs}
will be used to detect and remove paralogs prior to genotype calling.}

\item{filter_paralogs_alpha}{numeric, default 0.0016. Sites with p-values
below this will be considered paralogs if \code{filter_paralogs} is TRUE.
The default corresponds to a likelihood ratio score of roughly 10.}

\item{filter_paralogs_buffer}{numeric, default 1000. Sites within this many
bp of and candidate paralog sites will be excluded during genotyping if
\code{filter_paralogs} is TRUE.}

\item{filter_paralogs_populations}{character, default rep("only_pop",
length(bamfiles)). Vector of population names for each input bamfile for
use during paralog filtering. Loci will be considered candidate paralogs if
they are below \code{filter_paralogs_alpha} in any population.}

\item{filter_paralogs_reference}{character, default FALSE. If
\code{filter_paralogs} is TRUE, a reference genome is needed for the
creation of the target region .rf file needed by \code{SAMtools}.}

\item{rf}{logical, default FALSE. If TRUE, a path target region (.rf) file
containing genomic regions where loci will be sequenced. An rf file can be
supplied while still doing paralog filtering with if \code{filter_paralogs}
is TRUE, in which case only loci from both non-paralog regions and noted in
the rf file will be genotyped. Format: one region per line, noted by
'chrname:start-end'. The ending position can be omitted if the rest of the
remaining chromosome is a target region ('chr1:4124-' would keep every site
after position 4123). A single base can be specified using 'chrname:bp-bp'
(for example, 'chr1:4124-4124 would sequence only position 4124 on chr1).
Multiple regions per chromosome are acceptable. An entire chromosome can be
targeted by either omitting both the start and end or by noting both the
first and last bp ('chr1:1-', 'chr1:', and 'chr1:1-10000' would all include
the entire chr1 if it is 10000 bp long).}

\item{par}{numeric, default 1. Number of cores to allow ANGSD to use for
genotyping.}
}
\value{
Generates genotype files in the requested format named
  outfile_prefix.geno(.gz) in the directory of the first bamfile. The file
  will be zipped (.gz) if \code{unzip} is not TRUE. In R, returns file path
  to genotype output.
}
\description{
Uses the ANGSD (Analysis of Next Gen Sequencing Data) software to genotype
individuals across a set of bamfiles using several different filtering
options or genotyping methods. This function is a \emph{wrapper for ANGSD}.
}
\author{
William Hemstrom

Michael Miller
}
