% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATK.R
\name{run_HaplotypeCaller}
\alias{run_HaplotypeCaller}
\alias{add_RGS}
\alias{run_GenomicsDBImport}
\alias{run_GenotypeGVCFs}
\alias{run_VariantFiltration}
\alias{prep_genome_GATK}
\alias{genotype_bams_GATK}
\title{Genotype bamfiles with GATK}
\usage{
run_HaplotypeCaller(
  bamfiles,
  reference,
  mem,
  par = 1,
  java_path = "java",
  gatk4_path = "gatk.jar"
)

add_RGS(
  bamfiles,
  fastqs,
  par = 1,
  platform = "ILLUMINA",
  java_path = "java",
  picard_path = "picard.jar"
)

run_GenomicsDBImport(
  hapmap,
  bedfiles,
  mem,
  par = 1,
  batch_size = 5,
  java_path = "java",
  gatk4_path = "gatk.jar"
)

run_GenotypeGVCFs(
  bedfiles,
  reference,
  mem,
  par = 1,
  java_path = "java",
  gatk4_path = "gatk.jar"
)

run_VariantFiltration(
  vcfs,
  reference,
  mem,
  par = 1,
  QD = 2,
  FS = 60,
  SOR = 3,
  MQ = 40,
  MQRankSum = -12.5,
  ReadPosRankSum = -8,
  min_genotype_quality = 13,
  java_path = "java",
  gatk4_path = "gatk.jar"
)

prep_genome_GATK(reference, java_path = "java", picard_path = "picard.jar")

genotype_bams_GATK(
  bamfiles,
  reference,
  fastqs = NULL,
  par = 1,
  min_chr_size = 0,
  chunk_size = "chr",
  QD = 2,
  FS = 60,
  SOR = 3,
  MQ = 40,
  MQRankSum = -12.5,
  ReadPosRankSum = -8,
  min_genotype_quality = 13,
  platform = "ILLUMINA",
  batch_size = 5,
  add_RGs = TRUE,
  java_path = "java",
  gatk4_path = "gatk.jar",
  picard_path = "picard.jar",
  concatenate_final_vcfs = TRUE
)
}
\arguments{
\item{bamfiles}{Character, vector of bamfile paths. Bamfiles must be indexed,
as by default if produced using \code{\link{align_denovo}} or
\code{\link{align_reference}}.}

\item{reference}{Character, path to a reference genome. Must be indexed
with \code{bwa} and \code{picard} and zipped with \code{samtools bgzip}.
If running with \code{genotype_bams_GATK}, this will happen
automatically, if not, consider running \code{prep_genome_GATK}.}

\item{par}{Numeric, default 1. Number of cores to use for parallel processing.}

\item{java_path}{Character, default "java". Path to java install, by default
expects java on system path.}

\item{gatk4_path}{Character, default "gatk.jar". Path to gatk4 ".jar" file.
Used over a GATK direct system path install to allow for manual memory
allocation.}

\item{fastqs}{Character, paths to fastq files in the same order as bamfiles
(if required); for \code{add_RGs}. Not required if \code{add_RGs} is
\code{FALSE}.}

\item{platform}{Character, default "ILLUMINA". The platform to mark for
read-groups, ignored if \code{add_RGs} is \code{FALSE} for
\code{genotype_bams_GATK}.}

\item{picard_path}{Character, default "gatk.jar". Path to picard ".jar" file.}

\item{batch_size}{Numeric, default 5. The number of "gvcf" files to import at
a time during \code{GenomicsDBImport} operation. Higher numbers are faster
but more memory intense, for the most part.}

\item{QD}{Numeric, default 2. Minimum quality-by-depth score to keep
variants. See
\href{https://gatk.broadinstitute.org/hc/en-us/articles/360035890471-Hard-filtering-germline-short-variants}{here}
for details.}

\item{FS}{Numeric, default 60. Fisher strand bias score above which variants
will filtered out. See
\href{https://gatk.broadinstitute.org/hc/en-us/articles/360035890471-Hard-filtering-germline-short-variants}{here}
for details.}

\item{SOR}{Numeric, default 3. Strand odds-ratio test score test value above
which variants will be filtered out. See
\href{https://gatk.broadinstitute.org/hc/en-us/articles/360035890471-Hard-filtering-germline-short-variants}{here}
for details.}

\item{MQ}{Numeric, default 40. Mapping quality score below which variants
will be filtered out. See
\href{https://gatk.broadinstitute.org/hc/en-us/articles/360035890471-Hard-filtering-germline-short-variants}{here}
for details.}

\item{MQRankSum}{Numeric, default -12.5. Mapping quality rank-sum test score
below which variants will be filtered out. More negative values indicate
that the reads with the reference alleles map better than the alternative,
positive values indicate the inverse. See
\href{https://gatk.broadinstitute.org/hc/en-us/articles/360035890471-Hard-filtering-germline-short-variants}{here}
for details.}

\item{ReadPosRankSum}{Numeric, default -8. Minimum read position rank-sum test
score, below which variants will be filtered out. Measures the likelihood
that alternative alleles are more likely to appear farther along in reads
than the reference alleles. See
\href{https://gatk.broadinstitute.org/hc/en-us/articles/360035890471-Hard-filtering-germline-short-variants}{here}
for details.}

\item{min_genotype_quality}{Numeric, default 13. Phred-scaled genotype
quality score below which individual genotypes will be marked as missing.
The default corresponds to ~95% certainty of correct genotyping, which is
usually in the approximately 7x coverage range.}

\item{min_chr_size}{Numeric, default 0. Minimum size for chromosomes/scaffolds
for genotype calling. Useful to speed up computation by excluding small
unplaced scaffolds.}

\item{chunk_size}{Numeric or "chr", default "chr". Size of chunks by which
to split up databases and genotyping, in base pairs. If "chr", chromosomes
will not be split.}

\item{add_RGs}{Logical, default \code{TRUE}. If \code{TRUE}, basic read-group
information will be added to bam files using the function \code{add_RGs}.
If more complicated read-group information is already present in bam files,
this should be set to \code{FALSE}.}

\item{concatenate_final_vcfs}{Character, default \code{TRUE}. If \code{TRUE},
the final "vcf" files produced by \code{GenotypeGVCFs} will be concatenated
automatically using \code{\link{concat_vcfs}}.}
}
\value{
A vector of final ".vcf"/".gvcf"/"_db"/".bam"/.etc file paths,
  depending on the function function.
}
\description{
Run a HaplotypeCaller, GenomicsDBI, GenotypeGVCFs, VariantFiltration
pipeline to genotype and hardfilter bamfiles with GATK. GATK is much more
suitable for higher-coverage whole-genome sequencing data than ANGSD. Can
have \emph{very} long run-times.
}
\details{
Genotyping with GATK is a multistep, relatively complex process.
\code{genotype_bams_GATK} implements a pipeline to run through the major
steps in genotyping and filtering starting from aligned bamfiles, which can
be generated via \code{\link{align_denovo}} or \code{\link{align_reference}}.
This pipeline is \emph{much} slower than genotyping via
\code{\link{genotype_bams_ANGSD}}, but generally performs better for
high-coverage (7x+) data, particularly high-coverage whole-genome sequencing
data.

For more fine-grain parallel control, the individual component functions
\code{prep_genome_GATK}, \code{add_RGS}, \code{run_HaplotypeCaller},
\code{run_GenomicsDBImport}, \code{run_GenotypeGVCFs}, and
\code{run_VariantFiltration} can be run piecemeal. Note that \code{add_RGs}
adds basic read-group information assuming each bamfile has a unique read
group from the same library--manually adding RGs with \code{picard}`'s
\code{AddOrReplaceReadGroups} utility might be preferable if more detailed
RG info is available, in which case the \code{add_RGs} argument should be set
to \code{FALSE}.
}
\section{Functions}{
\itemize{
\item \code{run_HaplotypeCaller()}: Build ".gvcf" haplotype files for each sample
with \code{GATK}'s \code{HaplotypeCaller}.

\item \code{add_RGS()}: Add read groups to bam files with
\code{picard}.

\item \code{run_GenomicsDBImport()}: Build genome databases using \code{GATK}'s
\code{GenomicsDBImport}.

\item \code{run_GenotypeGVCFs()}: Generate called genotypes from GenomeDBs using
\code{GATK}'s \code{GenotypeGVCFs}.

\item \code{run_VariantFiltration()}: Filter called genotypes using \code{GATK}'s
\code{VariantFiltration}.

\item \code{prep_genome_GATK()}: Prepare a reference genome for genotyping with
\code{picard} and other indexers.

}}
\author{
William Hemstrom
}
