% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomics.R
\name{GetTCRs}
\alias{GetTCRs}
\title{Code to pull the full length TCR sequence from 10X genomics all_contig_annotations.json file}
\usage{
GetTCRs(
  barcodes,
  contig.annotations,
  json.path = ".",
  save.dir = ".",
  score = "",
  ranked.sheet = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{barcodes}{Barcodes used to identify cells in full contig annotations JSON file.}

\item{contig.annotations}{contig_annotations.csv file.}

\item{json.path}{Directory where JSON file is saved.}

\item{save.dir}{Directory where full length TCR transcripts will be saved.}

\item{score}{Orders the cells be classification score.}

\item{ranked.sheet}{False. If true implements and saves the cells by their characteristic rank.}

\item{verbose}{Print progress bars and output}
}
\value{
Void
}
\description{
This function will import a set of single cell barcodes, extract the full length TCR transcript details
and output the details ranked by the cell with the greatest functional score.
}
\concept{Genomics}
