% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{install_dependencies_conda}
\alias{install_dependencies_conda}
\title{Prepare a conda environment with `alignR`'s dependencies}
\usage{
install_dependencies_conda(name = "alignR")
}
\arguments{
\item{name}{character, default "alignR". The name of the conda env to create.}
}
\description{
Create a conda environment which has all of `alignR`'s main dependencies
installed save `alignR` itself which can be installed by activating the
env and installing `alignR` as usual.
}
\examples{
\dontrun{
# prepare environment
install_dependencies_conda("alignR")

# in terminal:
conda activate alignR
R -e "install.packages('remotes', repos = 'http://cran.us.r-project.org')"
R -e "remotes::install_github('hemstrow/alignR')"

}
}
\author{
William Hemstrom
}
