% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_Functions.R
\name{loadLMs}
\alias{loadLMs}
\title{Load Landmarks}
\usage{
loadLMs(lm_data, current_sp, lm_labels)
}
\arguments{
\item{lm_data}{An array of landmark data.}

\item{current_sp}{The current specimen, referencing the reactive object cur_sp().}

\item{lm_labels}{The landmark labels, referencing the reactive object LM_values().}
}
\value{
landmarks An array of loaded landmark data.
}
\description{
Loading of landmark data.
}
\details{
An internal function for loading of landmarks data. Used by various observe events in the shiny server,
including changing specimens and using the load button.
}
\keyword{internal}
