% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{locate_example_data}
\alias{locate_example_data}
\title{Locate the example reads and reference}
\usage{
locate_example_data()
}
\description{
Fetch a named list pointing to the RA, RB, and reference genome example 
files.
}
\examples{
# fetch the example data paths
example_data <- locate_example_data()
example_data$RA
example_data$RB
example_data$reference

}
