% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATK.R
\name{make_region_beds}
\alias{make_region_beds}
\title{Make .bed files for chromosomes or genomic chunks.}
\usage{
make_region_beds(reference, min_chr_size = 0, chunk_size = "chr", outdir = ".")
}
\arguments{
\item{reference}{Character. Path to reference genome.}

\item{min_chr_size}{Numeric, default 0. Minimum size for chromosomes/scaffolds.
Scaffolds smaller than this will be skipped when writing ".bed" files or
chunking the genome.}

\item{chunk_size}{Numeric or "chr", default "chr". Size of chunks by which
to split and write ".bed" files, in base pairs. If "chr", a ".bed" file
will be written for each chromosome/scaffold.}

\item{outdir}{Character, default ".". Path to directory where ".bed" files
will be written.}
}
\value{
A vector of written ".bed" file paths.
}
\description{
Generate ".bed" files describing eithe chromosomes or for automatically
tabulated chunks of the genome of a given size.
}
\author{
William Hemstrom
}
