% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomics.R
\name{MakeReceptor}
\alias{MakeReceptor}
\title{This function will import a set of single cell barcodes, extract the full length TCR transcript details
and output the details ranked by the cell with the greatest functional score.}
\usage{
MakeReceptor(n.cells, path = ".", save.dir = ".")
}
\arguments{
\item{n.cells}{Number of cells processed.}

\item{path}{Path to output files from alignr::GetTCRs}

\item{save.dir}{Directory where full length receptor sequences will be saved.}

\item{verbose}{Print progress bars and output}
}
\value{
Void
}
\description{
This function will import a set of single cell barcodes, extract the full length TCR transcript details
and output the details ranked by the cell with the greatest functional score.
}
\concept{Genomics}
