% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{merge_fastqs}
\alias{merge_fastqs}
\title{Merge fastq files}
\usage{
merge_fastqs(file_list, names = NULL, par = 1)
}
\arguments{
\item{file_list}{list where each element contains a set of fastq files to
merge.}

\item{names}{character vector, default NULL. A vector of names for each
merged output file. If not provided, files will be named "merged_x", where
x is the name of the first file in each set to merge.}

\item{par}{numeric, default 1. Number of cores to use for the merges.}
}
\value{
Generates merged fastq files. File paths provided as a character
  vector returned from this function.
}
\description{
Merges fastq files as a set from each element of a list. Resulting merged
files will be in the directory of the first fastq file of the first set to
merge.
}
\author{
William Hemstrom
}
