% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_Functions.R
\name{MeshManager}
\alias{MeshManager}
\title{Mesh Property Manager}
\usage{
MeshManager(object, color = "gray", size = 1, center = FALSE)
}
\arguments{
\item{object}{A mesh object as output by vcgImport or a matrix of values}

\item{size}{Numeric value which determines the size of vertex points for plot3d}

\item{center}{A logical value of whether the specimen should be centered or not}
}
\description{
Create a properly structured mesh object for landmarking in alignR.
}
\details{
This function pre-processes the mesh files for use in alignR, including determining the file type,
if the file should be centered, and checking coloring of the mesh. The output is a list with a
speicmen ID, mesh object, and point size values. This function is based off of
the code in digit.fixed from Geomorph 3.0.0.
}
