% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prep_scatter_intervals}
\alias{prep_scatter_intervals}
\title{Create scatter groups for \code{\link{run_rmake_pipeline}}.}
\usage{
prep_scatter_intervals(reference, chr_size_cutoff, max_chunk_size)
}
\arguments{
\item{reference}{character. Path to reference genome.}

\item{max_chunk_size}{numeric. The maximum size of any one genomic chunk
analyzed.}

\item{chr_size_cuttoff}{numeric. Chromosomes/scaffolds above this size
are considered as "chromosomes" for the purposes of scattering.}

\item{L}{length of the chromosome}

\item{S}{the starting value}

\item{s}{the number of bases in each segment (i.e. the seg_length)}

\item{m}{the desired max chunk size}
}
\value{
returns a tibble with start and stop columns of the different
chunks. (base-1 indexed, inclusive).
}
\description{
Generates scatter metadata for \code{\link{run_rmake_pipeline}} for use
in heavily parallel genotyping. Scaffolds are grouped by length, chromosomes
are split into chunks.
}
\author{
Eric Anderson

William Hemstrom
}
