% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomics.R
\name{RankTCRs}
\alias{RankTCRs}
\title{This function will import a set of single cell barcodes, extract the full length TCR transcript details
and output the details ranked by the cell with the greatest functional score.}
\usage{
RankTCRs(
  cell.data,
  gene = "IFNG",
  meta.data = FALSE,
  data.name = "Mdist",
  save.dir = ".",
  verbose = TRUE
)
}
\arguments{
\item{cell.data}{single cell RNA sequencing data processed by aocseq.}

\item{gene}{gene list for UMI ranked cell types.}

\item{meta.data}{select a metadata column to rank cell types.}

\item{data.name}{name of metadata used for ranking.}

\item{verbose}{Print progress bars and output.}
}
\value{
Void
}
\description{
This function will import a set of single cell barcodes, extract the full length TCR transcript details
and output the details ranked by the cell with the greatest functional score.
}
\concept{Genomics}
