% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmake.R
\name{run_rmake_pipeline}
\alias{run_rmake_pipeline}
\title{Run a complete GATK reads-to-genomes pipeline with rmake.}
\usage{
run_rmake_pipeline(
  RA,
  RB = NULL,
  read_metadata,
  reference,
  chr_size_cutoff,
  max_chunk_size,
  merge_list = NULL,
  until = NULL,
  downsample_args = list(low_cut = 0, high_cut = Inf),
  fastp_args = list(adapter_r1 = "AGATCGGAAGAGCACACGTCTGAACTCCAGTCA", adapter_r2 =
    "AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT", cut_front = FALSE, cut_front_mean_quality =
    NULL, cut_tail = FALSE, cut_tail_mean_quality = NULL, cut_right = 4,
    cut_right_mean_quality = 20, trim_poly_g = 10, trim_poly_x = FALSE),
  align_args = list(mapQ = 5, remove_duplicates = TRUE, remove_improper_pairs = TRUE,
    platform = "ILLUMINA"),
  HaplotypeCaller_args = list(mem = 2, min_base_quality_score = 33,
    minimum_mapping_quality = 20),
  GenomicsDBImport_args = list(mem = 4, batch_size = 5),
  GenotypeGVCFs_args = list(mem = 4, max_alternate_alleles = 2, new_qual = TRUE),
  VariantFiltration_args = list(mem = 2, QD = 2, FS = 60, SOR = 3, MQ = 40, MQRankSum =
    -12.5, ReadPosRankSum = -8, min_genotype_quality = 13),
  slurm_profile = list(genome_index = NULL, trim = NULL, align = NULL, merge = NULL,
    downsample = NULL, HaplotypeCaller = NULL, GenomicsDBImport = NULL, GenotypeGVCFs =
    NULL, VariantFiltration = NULL),
  slurm_system_append = NULL
)
}
\arguments{
\item{read_metadata}{a data.frame or equivalent with a row for each \code{RA}
read file and a columns, in order, indicating:

1. library
2. sampleID/name
3. sequencing platform
4. flowcell
5. sequencing lane

Column names are optional but recommended for organizational purposes.
The 'flowcell' and 'sequencing lane' columns can be left as NA,
in which case they will be guessed from the fastq header. This may create
errors if fastq headers are non-standard (flow cell and lane are expected
in the third and fourth slots, respectively (separated by ':')).}
}
\description{
Runs a reference-guided pipeline to go from demultiplexed reads to a vcf
file with called variants, including filtering. Optionally works with a slurm
system to queue jobs.
}
