% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_Functions.R
\name{saveLMs}
\alias{saveLMs}
\title{Save Landmarks}
\usage{
saveLMs(lm_data, current_sp, current_lm, lm_labels, keep)
}
\arguments{
\item{lm_data}{An array of landmark data.}

\item{current_sp}{The current specimen, referencing the reactive object cur_sp().}

\item{current_lm}{The current landmark, referencing the reactive object current_lm().}

\item{lm_labels}{The landmark labels, referencing the reactive object LM_values().}

\item{keep}{The newly confirmed landmark data.}
}
\value{
landmarks An array of updated landmark data.
}
\description{
Saving or writing landmark data.
}
\details{
An internal function for saving new landmarks when confirmed. Used by confirmLM() in the shiny server.
}
\keyword{internal}
