% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPan3d.R
\name{shinyPan3d}
\alias{shinyPan3d}
\title{shinyPan3d}
\usage{
shinyPan3d(par_input, shinyBrush, session)
}
\arguments{
\item{par_input}{An input from shinyGetPar3d & alignRPar3d to get the model, projection, and view for the current orientation of the mesh in the rglwidget.}

\item{shinyBrush}{The ID for the shinyBrush shared from the rglwidget which contains the output of the click}
}
\description{
This function uses the shinyBrush to locate position of click for moving specimen in rglwidget.
}
\details{
This internal function
}
