% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_Functions.R
\name{shinySelectPoints3d}
\alias{shinySelectPoints3d}
\title{shinySelectPoints3d}
\usage{
shinySelectPoints3d(centers, verts, tris, N, par_input, shinyBrush)
}
\arguments{
\item{centers}{An array of the center points for each triangle.}

\item{verts}{An array of triangle vertices for the mesh.}

\item{tris}{A matrix containing vertex indices of the triangles of the mesh.}

\item{N}{The number of triangles.}

\item{par_input}{An input from shinyGetPar3d & alignRPar3d to get the model, projection, and view for the current orientation of the mesh in the rglwidget.}

\item{shinyBrush}{The ID for the shinyBrush shared from the rglwidget which contains the output of the click}
}
\description{
This function identifies where a click on the rglwidget screen is positioned on the surface object.
}
\details{
This internal function takes the mesh values, rgl parameters, and shinyBrush values for
a click input and identifies the position of the click on the surface mesh. MORE DETAILS?
}
