% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sort_bams}
\alias{sort_bams}
\title{Sort bam files}
\usage{
sort_bams(bams, par = 1)
}
\arguments{
\item{bams}{A list of bamfiles to be sorted.}

\item{par}{numeric, default 1. Number of cores to use for sorting.}
}
\value{
Generates sorted bam files. File paths provided as a character
  vector returned from this function.
}
\description{
Sorts bam files using samtools.
}
\author{
William Hemstrom
}
