% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{trim_fastp}
\alias{trim_fastp}
\title{Trim reads with fastp}
\usage{
trim_fastp(
  RA_fastqs,
  RB_fastqs = NULL,
  adapter_r1 = "AGATCGGAAGAGCACACGTCTGAACTCCAGTCA",
  adapter_r2 = "AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT",
  cut_front = FALSE,
  cut_front_mean_quality = NULL,
  cut_tail = FALSE,
  cut_tail_mean_quality = NULL,
  cut_right = 4,
  cut_right_mean_quality = 20,
  trim_poly_g = 10,
  trim_poly_x = FALSE,
  par = 1
)
}
\arguments{
\item{RA_fastqs}{character. Vector of filepaths for the RA (read one) files.}

\item{RB_fastqs}{character or NULL, default NULL. Vector of filepaths for the
RB (read two) files.}

\item{adapter_r1}{character or NULL, default
"AGATCGGAAGAGCACACGTCTGAACTCCAGTCA". Adapter sequences for read 1,
automatically detected by fastp if NULL. The defaults are the TruSeq
adapters, which are far and away the most common.}

\item{adapter_r2}{character or NULL, default
"AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT". Adapter sequences for read 1,
automatically detected by fastp if NULL. The defaults are the TruSeq
adapters, which are far and away the most common.}

\item{cut_front}{numeric or FALSE, default FALSE. If a number, a sliding
window size by which to scan (starting from the 5' end) for average quality
windows below `cut_front_mean_quality`. Bases will be dropped until a
window above `cut_front_mean_quality` is detected. NOTE: this will
interfere with duplicate detection and removal during later stages.}

\item{cut_front_mean_quality}{numeric or NULL, default NULL. If a number,
sliding windows of size `cut_front` will be scanned (starting from the 5'
end) for average quality windows below `cut_front_mean_quality`. Bases will
be dropped until a window above `cut_front_mean_quality` is detected. NOTE:
this will interfere with duplicate detection and removal during later
stages.}

\item{cut_tail}{numeric or FALSE, default FALSE. If a number, a sliding
window size by which to scan (starting from the 3' end) for average quality
windows below `cut_front_mean_quality`. Bases will be dropped until a
window above `cut_front_mean_quality` is detected.}

\item{cut_tail_mean_quality}{numeric or NULL, default NULL. If a number,
sliding windows of size `cut_tail` will be scanned (starting from the 3'
end) for average quality windows below `cut_tail_mean_quality`. Bases will
be dropped until a window above `cut_tail_mean_quality` is detected.}

\item{cut_right}{numeric or FALSE, default 4. If a number, a sliding
window size by which to scan for average quality windows below
`cut_front_right_quality`. All bases to the right of any bad window will be
removed.}

\item{trim_poly_g}{numeric or FALSE, default 10. If TRUE, polyG tails will be
detected and removed if they are longer that `trim_poly_g`.}

\item{trim_poly_x}{numeric or FALSE, default FALSE. If a number, polyX (A, T,
C, G) tails will be detected and removed if they are longer than
`trim_poly_x`. Detection happens after polyG tail removal if `trim_poly_g`
is also true}

\item{par}{numeric, default 1. Number of cores to use for the alignments.}

\item{cut_front_right_quality}{numeric or NULL, default 20. If a number,
sliding windows of size `cut_right` will be scanned for average quality
windows below `cut_front_right_quality`. All bases to the right of any bad
window will be removed.}
}
\value{
Generates 'x_trimmed.fastq' files (sorted
  alignments and index files, respectively), where 'x' is the prefix for the
  RA/RB fastqs. In R, returns a vector of output file paths.
}
\description{
Trim reads with fastp
}
\author{
William Hemstrom
}
