% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_denovo}
\alias{align_denovo}
\title{Filter and align single or paired-end to a denovo assembly via stacks.}
\usage{
align_denovo(
  RA_fastqs,
  RB_fastqs = NULL,
  M,
  n = M,
  par = 1,
  check_headers = FALSE,
  stacks_cleanup = TRUE,
  re_align = TRUE,
  mapQ = 5,
  remove_duplicates = TRUE,
  remove_improper_pairs = TRUE,
  ask_confirmation = TRUE
)
}
\arguments{
\item{RA_fastqs}{character. Vector of filepaths for the RA (read one) files.}

\item{RB_fastqs}{character or NULL, default NULL. Vector of filepaths for the
RB (read one) files.}

\item{M}{numeric. The \code{-M} parameter from STACKS, the "number of
mismatches allowed between stacks within individuals" (see STACKS
documentation).}

\item{n}{numeric, default \code{M}. The \code{-n} parameter from STACKS, the
"number of mismatches allowed between stacks between individuals" (see
STACKS documentation).}

\item{par}{numeric, default 1. Number of cores to use for the alignments.}

\item{check_headers}{logical, default TRUE. If TRUE, will check that fastq
headers are reasonable (end in /1 and /2 and otherwise match between paired
ends) before running STACKS if doing denovo assembly with paired-end reads.
Somewhat computationally and I/O intensive, so disable if confident that
headers are OK.}

\item{stacks_cleanup}{logical, default TRUE. If TRUE, files other than .bam
files created by stacks will be removed.}

\item{re_align}{logical, default TRUE. If TRUE, bam files will be re-aligned
to the denovo reference generated by STACKS using \code{bwa mem} via
\code{align_reference}. This is useful if "truncated file" errors are
generated during indexing, which can occasionally occur when stacks aligns
reads.}

\item{mapQ}{numeric, default 5. Used only if re_align = TRUE, passed to
\code{align_reference}. Reads with mapping qualities lower than mapQ will
be removed using \code{samtools view -q mapQ}. Only valid for paired-end
reads.}

\item{remove_duplicates}{logical, default TRUE. Used only if re_align = TRUE,
  passed to \code{align_reference}. If TRUE, PCR duplicates (clones) will be
removed using \code{samtools markdup}. Only valid for paired-end reads.}

\item{remove_improper_pairs}{logical, default FALSE. Used only if re_align =
TRUE, passed to \code{align_reference}. If TRUE, improperly paired reads
reads will be removed using \code{samtools view -f 0x2}. Only valid for
paired-end reads, and should usually be FALSE for denovo alignments.}

\item{ask_confirmation}{logical, default TRUE. If TRUE, \code{alignR} will
ask for confirmation before moving or renaming files. If FALSE,
\code{alignR} will move or rename without confirmation.}
}
\value{
If the \code{stacks_cleanup} arugment is TRUE, catalog.fa and index
  files for the denovo genome in the directory containing the fastq files
  alongside alignments, either 'x.alns.bam', 'x.alns.sort.bam', and
  'x.alns.sort.bam.bai' or 'x.sort.flt.bam' and 'x.sort.flt.bam.bai' where
  'x' is the prefix for the RA fastqs, depending on if re-alignment with
  \code{bwa mem} via \code{align_reference} is requested with the
  \code{re_align} argument. In R, returns a vector of output file paths.
}
\description{
Aligns split, single- or paired-end reads to a reference genome using the
stacks toolkit. This script is a wrapper for a slightly edited
\code{denovo_map.pl} that is adjusted to not use \code{gstacks} or
\code{populations}, the former of which can be run with
\code{\link{genotype_bams}}. The functionality of the later of which can be
done using several other R packages.
}
\details{
Paired-end reads are filtered following alignment by first removing PCR
duplicates using \code{samtools fixmate} and \code{samtools markdup}, then
poorly mapped reads using \code{samtools view} with \code{-q 5}, and then
improperly paired reads using \code{samtools view} with  \code{-f 0x2}. The
full alignment script can be accessed using \code{system.file("shell",
"run_align.sh", "alignR")}.

Single-end reads are not filtered following alignment, although filtering can
(and usually should) be done while calling genotypes or doing other
down-stream analyses. The full alignment script can be accessed using
\code{system.file("shell", "run_align_single.sh", "alignR")}.
}
\author{
William Hemstrom
}
