% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Split multiplexed reads with sample barcodes.}
\usage{
demultiplex(
  R1_files,
  R2_files,
  barcodes,
  outfile_prefix = "alignR_",
  sample_names = NULL
)
}
\arguments{
\item{barcodes}{character. Vector of barcodes to search for in read files.}

\item{outfile_prefix}{character. prefix to be appeneded to each resulting
.fastq file.}

\item{sample_names}{character. Vector of names for output files corresponding
in order to the provided barcodes. Prefix and barcodes will in names
will be replaced with these names.}

\item{R1}{character. File name for R1 (read one) file.}

\item{R2}{character. File name for R2 (read one) file.}
}
\description{
Splits paired-end sequencing .fastq files containing reads from multiple
different plates with unique plate barcodes into different fastq files for
each barcode. .fastq files must have barcode identifiers starting at the 5th
base of sequence data.
}
\details{
Currently, this only supports "strict" barcode matching without *any*
sequence mismatches.
}
\author{
William Hemstrom

Michael Miller
}
