% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{plate_split}
\alias{plate_split}
\title{Split paired-end .fastq files with reads from different plates.}
\usage{
plate_split(R1, R2 = NULL, R3, indices, outfile_prefix = "plate_split")
}
\arguments{
\item{R1}{character. File name for R1 (read one) file.}

\item{R2}{character or NULL, default NULL. File name for R2 (index read)
  file. If NULL, if NULL assumes that the plate index is at the end of the
.fastq headers the R1 and R3 files.}

\item{R3}{character. File name for R3 (read two) file}

\item{indices}{character. Vector of indices to search for in read files.}

\item{outfile_prefix}{character, default plate_split. prefix to be appended
to each resulting .fastq file.}
}
\value{
Generates files split by index in the directory of the R1 file(s),
  named outfile_prefix_RN_INDEX.fastq, where N is 1, 2, 3 three (for the R1,
  R2, and R3 files). In R, returns a list containing the paths to each output
  file.
}
\description{
Splits paired-end sequencing .fastq files containing reads from multiple
different plates with unique plate indices into different fastq files for
each plate index.
}
\details{
The R1 and R3 .fastq files should contain the actual forward and reverse
reads. The optional R2 .fastq file should contain only the plate indices. As
is standard, the reads are assumed to be in the same order for each of the
.fastq files. If no R2 file is provided, the index is assumed to be at the
end of the headers for each read in the R1 and R3 files.
}
\author{
William Hemstrom

Michael Miller
}
