% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{plate_split}
\alias{plate_split}
\title{Split paired-end .fastq files with reads from different plates.}
\usage{
plate_split(R1, R2, R3, barcodes, outfile_prefix = "")
}
\arguments{
\item{R1}{character. File name for R1 (read one) file.}

\item{R2}{character. File name for R2 (read one) file.}

\item{barcodes}{character. Vector of barcodes to search for in read files.}

\item{outfile_prefix}{character. prefix to be appeneded to each resulting
.fastq file.}
}
\description{
Splits paired-end sequencing .fastq files containing reads from multiple
different plates with unique plate barcodes into different fastq files for
each barcode. .fastq files must have barcode identifiers at the start of the
second line for each read (the first line of information).
}
\author{
William Hemstrom

Michael Miller
}
